+++ core/domain/src/main/java/com/godlife/domain/GetMostViewStimulusPostUseCase.kt
@@ -8,5 +8,5 @@ import javax.inject.Inject
 class GetMostViewStimulusPostUseCase @Inject constructor(
     private val networkRepository: NetworkRepository
 ) {
-    suspend fun executeGetMostViewStimulusPost(authorization: String) = networkRepository.getStimulusMostViewPost(authorization)
+    suspend fun executeGetGetMostViewStimulusPost(authorization: String) = networkRepository.getStimulusMostViewPost(authorization)
 }
