+++ feature/community-page/src/main/java/com/godlife/community_page/stimulus/StimulusPostScreen.kt
@@ -18,7 +18,6 @@ import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
@@ -26,13 +25,9 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.LazyRow
 import androidx.compose.foundation.lazy.grid.GridCells
 import androidx.compose.foundation.lazy.grid.LazyHorizontalGrid
 import androidx.compose.foundation.lazy.grid.itemsIndexed
-import androidx.compose.foundation.lazy.itemsIndexed
-import androidx.compose.foundation.pager.HorizontalPager
-import androidx.compose.foundation.pager.rememberPagerState
 import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.rounded.Create
@@ -47,25 +42,20 @@ import androidx.compose.material3.Icon
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.MutableState
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableIntStateOf
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.blur
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.draw.shadow
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.asImageBitmap
 import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.layout.onGloballyPositioned
 import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.platform.LocalDensity
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.font.FontWeight
@@ -79,26 +69,25 @@ import androidx.navigation.NavController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
-import androidx.paging.compose.collectAsLazyPagingItems
 import com.bumptech.glide.Glide
 import com.bumptech.glide.request.target.CustomTarget
 import com.bumptech.glide.request.transition.Transition
 import com.godlife.community_page.BuildConfig
 import com.godlife.community_page.R
-import com.godlife.community_page.navigation.StimulusPostDetailRoute
 import com.godlife.community_page.search.stimulus.StimulusSearchScreen
+import com.godlife.community_page.stimulus.famous_post.FamousStimulusPostContent
+import com.godlife.community_page.stimulus.latest_post.LatestStimulusPostContent
+import com.godlife.community_page.stimulus.latest_post.LatestStimulusPostListPreview
+import com.godlife.community_page.stimulus.most_view_post.MostViewStimulusPostContent
+import com.godlife.community_page.stimulus.recommended_author_post.RecommendedAuthorInfoContentPreview
+import com.godlife.community_page.stimulus.recommended_author_post.RecommendedAuthorStimulusPostContent
 import com.godlife.community_page.stimulus.recommended_post.RecommendedStimulusPostContent
-import com.godlife.community_page.stimulus.recommended_post.RecommendedStimulusPostContentPreview
 import com.godlife.create_post.stimulus.CreateStimulusPostScreen
 import com.godlife.designsystem.theme.GodLifeTheme
 import com.godlife.designsystem.theme.GrayWhite
 import com.godlife.designsystem.theme.OpaqueDark
 import com.godlife.designsystem.theme.PurpleMain
 import com.godlife.network.model.StimulusPostList
-import kotlinx.coroutines.NonCancellable
-import kotlinx.coroutines.delay
-import kotlinx.coroutines.launch
-import kotlinx.coroutines.withContext
 
 data class FABItem(
     val icon: ImageVector,
@@ -184,7 +173,7 @@ fun StimulusPostScreen(
                         }
 
                         item {
-                            StimulusPostFamousContentPreview()
+                            FamousStimulusPostContent(navController = navController)
                         }
 
                         item {
@@ -202,9 +191,32 @@ fun StimulusPostScreen(
 
                         }
 
-                        item { LatestStimulusPostList(viewModel = viewModel, navController = navController) }
+                        item {
+                            LatestStimulusPostContent(
+                                navController = navController
+                            )
+                        }
 
-                        item { RecommendUserContentPreview() }
+                        item{
+                            Text(
+                                modifier = modifier
+                                    .padding(20.dp),
+                                text = "조회수가 높은 글이에요.",
+                                style = TextStyle(
+                                    color = Color.Black,
+                                    fontSize = 18.sp,
+                                    fontWeight = FontWeight.Bold
+                                )
+                            )
+                        }
+
+                        item{
+                            MostViewStimulusPostContent(
+                                navController = navController
+                            )
+                        }
+
+                        item { RecommendedAuthorInfoContentPreview() }
 
                         item {
 
@@ -221,7 +233,11 @@ fun StimulusPostScreen(
 
                         }
 
-                        item { LatestStimulusPostListPreview() }
+                        item {
+                            RecommendedAuthorStimulusPostContent(
+                                navController = navController
+                            )
+                        }
 
 
                     }
@@ -249,142 +265,6 @@ fun StimulusPostScreen(
 
 }
 
-@Composable
-fun LatestStimulusPostList(
-    modifier: Modifier = Modifier,
-    viewModel: StimulusPostViewModel,
-    navController: NavController
-){
-    val postList = viewModel.latestPostList.collectAsLazyPagingItems()
-
-    /*
-    items(postList.itemCount){
-        postList[it]?.let { it1 -> LatestPostListView(item = it1, navController = navController) }
-    }
-     */
-
-    LazyHorizontalGrid(
-        modifier = modifier
-            .fillMaxWidth()
-            .height(500.dp),
-        rows = GridCells.Fixed(3)
-    ) {
-
-        items(postList.itemCount){
-            postList[it]?.let { it1 -> LatestStimulusItem(item = it1, navController = navController)  }
-        }
-    }
-}
-
-@Composable
-fun LatestStimulusItem(
-    modifier: Modifier = Modifier,
-    item: StimulusPostList,
-    navController: NavController
-){
-
-    val postId = item.boardId.toString()
-
-
-    val bitmap: MutableState<Bitmap?> = remember { mutableStateOf(null) }
-
-    Row(
-        modifier = modifier
-            .size(height = 150.dp, width = 300.dp)
-            .background(Color.White)
-            .padding(10.dp)
-            .clickable { navController.navigate("${StimulusPostDetailRoute.route}/$postId") },
-        verticalAlignment = Alignment.CenterVertically
-    ){
-
-        Box(
-            modifier = modifier
-                .padding(10.dp)
-                .size(width = 100.dp, height = 150.dp)
-                .shadow(10.dp),
-            contentAlignment = Alignment.Center
-        ){
-
-            Glide.with(LocalContext.current)
-                .asBitmap()
-                .load(BuildConfig.SERVER_IMAGE_DOMAIN + item.thumbnailUrl)
-                .error(R.drawable.category3)
-                .into(object : CustomTarget<Bitmap>() {
-                    override fun onResourceReady(resource: Bitmap, transition: Transition<in Bitmap>?) {
-                        bitmap.value = resource
-                    }
-
-                    override fun onLoadCleared(placeholder: Drawable?) {}
-                })
-
-            bitmap.value?.asImageBitmap()?.let { fetchedBitmap ->
-                Image(
-                    bitmap = fetchedBitmap,
-                    contentDescription = null,
-                    contentScale = ContentScale.Crop,
-                    modifier = modifier
-                        .fillMaxWidth()
-                )   //bitmap이 없다면
-            } ?: Image(
-                painter = painterResource(id = R.drawable.category3),
-                contentDescription = null,
-                contentScale = ContentScale.Crop,
-                modifier = modifier
-                    .fillMaxWidth()
-            )
-
-            Box(
-                modifier = modifier
-                    .fillMaxSize()
-                    .padding(15.dp)
-                    .background(color = OpaqueDark)
-                    .padding(start = 10.dp, end = 10.dp, top = 10.dp, bottom = 10.dp),
-                contentAlignment = Alignment.Center
-            ){
-
-                Text(text = item.title,
-                    style = TextStyle(color = Color.White, fontSize = 12.sp, fontWeight = FontWeight.Bold),
-                    overflow = TextOverflow.Ellipsis
-                )
-
-            }
-
-        }
-
-        Spacer(modifier.size(10.dp))
-
-        Column(
-
-        ) {
-            Text(
-                text = item.introduction,
-                style = TextStyle(
-                    color = GrayWhite,
-                    fontSize = 12.sp,
-                    fontWeight = FontWeight.Normal,
-                ),
-                overflow = TextOverflow.Ellipsis
-            )
-
-            Spacer(modifier.size(5.dp))
-
-            HorizontalDivider()
-
-            Spacer(modifier.size(5.dp))
-
-            Text(
-                modifier = modifier
-                    .fillMaxWidth(),
-                text = "by ${item.nickname}",
-                style = TextStyle(
-                    color = GrayWhite,
-                    fontSize = 12.sp,
-                    fontWeight = FontWeight.Bold,
-                )
-            )
-        }
-    }
-}
 
 @Composable
 fun StimulusCoverItem(
@@ -585,395 +465,9 @@ fun StimulusCoverItemPreview(
     }
 }
 
-@Preview
-@Composable
-fun StimulusPostFamousContentPreview(
-    modifier: Modifier = Modifier,
-    navController: NavController? = null
-){
-    val item = listOf(
-        StimulusPostItem(
-            title = "이것은 제목!",
-            writer = "치킨 러버",
-            coverImg = R.drawable.category3,
-            introText = "대충 이러이러한 내용이라고 소개하는 내용"
-        ),
-        StimulusPostItem(
-            title = "나도 제목!",
-            writer = "초코 러버",
-            coverImg = R.drawable.category4,
-            introText = "대충 이러이러한 내용이라고 소개하는 내용"
-        ),
-        StimulusPostItem(
-            title = "제목 등장",
-            writer = "라면 좋아",
-            coverImg = R.drawable.category3,
-            introText = "대충 이러이러한 내용이라고 소개하는 내용"
-        ),
-        StimulusPostItem(
-            title = "제모오오오옥",
-            writer = "헬로우",
-            coverImg = R.drawable.category4,
-            introText = "대충 이러이러한 내용이라고 소개하는 내용"
-        )
-    )
-
-    LazyRow {
-        itemsIndexed(item){index, it ->
-
-            FamousStimulusItemPreview(item = it)
-
-        }
-    }
-}
-
-@Preview
-@Composable
-fun FamousStimulusItemPreview(
-    modifier: Modifier = Modifier,
-    item: StimulusPostItem = StimulusPostItem(title = "이것이 제목이다", writer = "치킨 러버", coverImg = R.drawable.category3, introText = "갓생을 살고 싶어하는 당신을 위해 작성한 글!")
-){
-    Column(
-        modifier = modifier
-            .size(width = 300.dp, height = 350.dp),
-        verticalArrangement = Arrangement.Center,
-        horizontalAlignment = Alignment.CenterHorizontally
-    ) {
-
-        Box(
-            modifier = modifier
-                .padding(10.dp)
-                .size(width = 200.dp, height = 250.dp)
-                .shadow(10.dp),
-            contentAlignment = Alignment.Center
-        ){
-            Image(
-                modifier = modifier
-                    .fillMaxWidth(),
-                painter = painterResource(id = item.coverImg),
-                contentDescription = "",
-                contentScale = ContentScale.Crop
-            )
-
-            Box(
-                modifier = modifier
-                    .fillMaxSize()
-                    .padding(30.dp)
-                    .background(color = OpaqueDark)
-                    .padding(start = 20.dp, end = 20.dp, top = 10.dp, bottom = 10.dp),
-                contentAlignment = Alignment.Center
-            ){
-
-                Text(text = item.title,
-                    style = TextStyle(color = Color.White, fontSize = 18.sp, fontWeight = FontWeight.Bold)
-                )
-
-            }
-
-        }
-
-        Spacer(modifier.size(5.dp))
-
-        Text(
-            text = item.introText,
-            style = TextStyle(
-                color = GrayWhite,
-                fontSize = 12.sp,
-                fontWeight = FontWeight.Normal,
-                textAlign = TextAlign.Center
-            )
-        )
-
-        Spacer(modifier.size(5.dp))
-
-        HorizontalDivider(modifier.padding(start = 20.dp, end = 20.dp))
-
-        Spacer(modifier.size(5.dp))
-
-        Text(
-            text = "by.${item.writer}",
-            style = TextStyle(
-                color = GrayWhite,
-                fontSize = 12.sp,
-                fontWeight = FontWeight.Normal,
-                textAlign = TextAlign.Center
-            )
-        )
-
-    }
-
-}
-
-@Preview
-@Composable
-fun LatestStimulusPostListPreview(
-    modifier: Modifier = Modifier
-){
-
-    val item = listOf(
-        StimulusPostItem(
-            title = "이것은 제목!",
-            writer = "치킨 러버",
-            coverImg = R.drawable.category3,
-            introText = "대충 이러이러한 내용이라고 소개하는 내용"
-        ),
-        StimulusPostItem(
-            title = "나도 제목!",
-            writer = "초코 러버",
-            coverImg = R.drawable.category4,
-            introText = "대충 이러이러한 내용이라고 소개하는 내용"
-        ),
-        StimulusPostItem(
-            title = "제목 등장",
-            writer = "라면 좋아",
-            coverImg = R.drawable.category3,
-            introText = "대충 이러이러한 내용이라고 소개하는 내용"
-        ),
-        StimulusPostItem(
-            title = "제모오오오옥",
-            writer = "헬로우",
-            coverImg = R.drawable.category4,
-            introText = "대충 이러이러한 내용이라고 소개하는 내용"
-        ),
-        StimulusPostItem(
-            title = "제목 등장",
-            writer = "라면 좋아",
-            coverImg = R.drawable.category3,
-            introText = "대충 이러이러한 내용이라고 소개하는 내용"
-        ),
-        StimulusPostItem(
-            title = "제모오오오옥",
-            writer = "헬로우",
-            coverImg = R.drawable.category4,
-            introText = "대충 이러이러한 내용이라고 소개하는 내용"
-        )
-    )
-
-    LazyHorizontalGrid(
-        modifier = modifier
-            .fillMaxWidth()
-            .height(500.dp),
-        rows = GridCells.Fixed(3)
-    ) {
-        itemsIndexed(item){index, it ->
-            LatestStimulusItemPreview(item = it)
-        }
-    }
-}
-
-@Preview
-@Composable
-fun LatestStimulusItemPreview(
-    modifier: Modifier = Modifier,
-    item: StimulusPostItem = StimulusPostItem(title = "이것이 제목이다", writer = "치킨 러버", coverImg = R.drawable.category3, introText = "갓생을 살고 싶어하는 당신을 위해 작성한 글!")
-){
-    Row(
-        modifier = modifier
-            .size(height = 150.dp, width = 300.dp)
-            .background(Color.White)
-            .padding(10.dp),
-        verticalAlignment = Alignment.CenterVertically
-    ){
-        Image(
-            modifier = modifier
-                .size(100.dp),
-            painter = painterResource(id = item.coverImg),
-            contentDescription = "",
-            contentScale = ContentScale.Crop
-        )
-
-        Spacer(modifier.size(10.dp))
-
-        Column(
-
-        ) {
-            Text(
-                text = item.introText,
-                style = TextStyle(
-                    color = GrayWhite,
-                    fontSize = 12.sp,
-                    fontWeight = FontWeight.Normal,
-                ),
-                overflow = TextOverflow.Ellipsis
-            )
-
-            Spacer(modifier.size(5.dp))
-
-            HorizontalDivider()
-
-            Spacer(modifier.size(5.dp))
-
-            Text(
-                modifier = modifier
-                    .fillMaxWidth(),
-                text = "by.${item.writer}",
-                style = TextStyle(
-                    color = GrayWhite,
-                    fontSize = 12.sp,
-                    fontWeight = FontWeight.Bold,
-                )
-            )
-        }
-    }
-}
-
-@Preview
-@Composable
-fun RecommendUserContentPreview(
-    modifier: Modifier = Modifier
-){
-
-    val user = RecommendUserItem(
-        nickname = "치킨 러버",
-        introText = "치킨을 좋아하는 사람입니다.",
-        profileImg = R.drawable.category3,
-        backgroundImg = R.drawable.category4,
-        postItem = listOf(
-            StimulusPostItem(
-                title = "이것은 제목!",
-                writer = "치킨 러버",
-                coverImg = R.drawable.category3,
-                introText = "대충 이러이러한 내용이라고 소개하는 내용"
-            ),
-            StimulusPostItem(
-                title = "나도 제목!",
-                writer = "초코 러버",
-                coverImg = R.drawable.category4,
-                introText = "대충 이러이러한 내용이라고 소개하는 내용"
-            ),
-            StimulusPostItem(
-                title = "제목 등장",
-                writer = "라면 좋아",
-                coverImg = R.drawable.category3,
-                introText = "대충 이러이러한 내용이라고 소개하는 내용"
-            ),
-            StimulusPostItem(
-                title = "제모오오오옥",
-                writer = "헬로우",
-                coverImg = R.drawable.category4,
-                introText = "대충 이러이러한 내용이라고 소개하는 내용"
-            )
-        )
-    )
-
-    Box(
-        modifier = modifier
-            .height(400.dp)
-            .fillMaxWidth()
-            .background(Color.Black),
-        contentAlignment = Alignment.Center
-    ){
-
-
-        Image(
-            modifier = modifier
-                .fillMaxSize()
-                .blur(
-                    radiusX = 15.dp, radiusY = 15.dp
-                ),
-            painter = painterResource(id = user.backgroundImg),
-            contentDescription = "",
-            contentScale = ContentScale.Crop,
-            alpha = 0.6f
-        )
-
-        RecommendUserProfilePreview()
-
-    }
-
-}
-
-@Preview
-@Composable
-fun RecommendUserProfilePreview(
-    modifier: Modifier = Modifier,
-    item: RecommendUserItem = RecommendUserItem(nickname = "치킨 러버", introText = "치킨을 좋아하는 사람입니다.", profileImg = R.drawable.category3, backgroundImg = R.drawable.category4, postItem = emptyList())
-){
-    Column(
-        modifier = modifier
-            .fillMaxSize()
-            .padding(20.dp),
-        horizontalAlignment = Alignment.CenterHorizontally,
-        verticalArrangement = Arrangement.Center
-    ) {
-
-        Text(
-            modifier = modifier
-                .fillMaxWidth(),
-            text = "추천 작가",
-            style = TextStyle(
-                color = Color.White,
-                fontSize = 18.sp,
-                fontWeight = FontWeight.Bold
-            ),
-            textAlign = TextAlign.Start
-        )
-
-        Spacer(modifier.size(20.dp))
-
-        Row(
-            modifier = modifier
-                .fillMaxSize()
-                .padding(10.dp),
-            verticalAlignment = Alignment.CenterVertically,
-            horizontalArrangement = Arrangement.Center
-        ){
-
-            Column(
-                modifier = modifier
-                    .weight(0.5f),
-                horizontalAlignment = Alignment.CenterHorizontally
-            ) {
-
-                Image(
-                    modifier = modifier
-                        .size(150.dp)
-                        .clip(shape = CircleShape),
-                    painter = painterResource(id = item.profileImg),
-                    contentDescription = "",
-                    contentScale = ContentScale.Crop
-                )
-
-                Spacer(modifier.size(20.dp))
-
-                Text(
-                    text = item.nickname,
-                    style = TextStyle(
-                        color = Color.White,
-                        fontSize = 18.sp,
-                        fontWeight = FontWeight.Bold
-                    ),
-                    textAlign = TextAlign.Center
-                )
-            }
-
-            Box(
-                modifier = modifier
-                    .weight(0.5f)
-            ){
-
-                Text(
-                    modifier = modifier
-                        .padding(start = 20.dp, end = 20.dp),
-                    text = "\"${item.introText}\"",
-                    style = TextStyle(
-                        color = Color.White,
-                        fontSize = 18.sp,
-                        fontWeight = FontWeight.Normal
-                    ),
-                    textAlign = TextAlign.Center
-                )
-
-            }
 
 
 
-        }
-
-
-    }
-
-}
 
 @Preview
 @Composable
@@ -983,7 +477,8 @@ fun StimulusLoadingScreen(
     Box(
         modifier = modifier
             .background(Color.White)
-            .fillMaxSize(),
+            .fillMaxWidth()
+            .height(300.dp),
         contentAlignment = Alignment.Center
     ){
         CircularProgressIndicator(color = PurpleMain)
